change sig_atomic_t from long to int matching all the other archs
ok kettenis@

+++ sys/arch/riscv64/include/signal.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: signal.h,v 1.2 2021/05/12 01:20:52 jsg Exp $	*/
+/*	$OpenBSD: signal.h,v 1.3 2021/05/13 08:44:18 jsg Exp $	*/
 
 /*-
  * Copyright (c) 1986, 1989, 1991, 1993
@@ -39,7 +39,7 @@
 
 #include <sys/cdefs.h>
 
-typedef	long sig_atomic_t;
+typedef int sig_atomic_t;
 
 #if __BSD_VISIBLE || __XPG_VISIBLE >= 420
